removing factory file
