+++ pkg/certificateprovider/inline/provider_test.go
@@ -57,8 +57,7 @@ func TestGetCertificates(t *testing.T) {
 		},
 	}
 
-	providerFactory := &inlineCertProviderFactory{}
-	provider, _ := providerFactory.Create()
+	provider := Create()
 
 	for _, tc := range cases {
 		t.Run(tc.desc, func(t *testing.T) {
